+++ docsig/_config.py
@@ -80,6 +80,12 @@ class Parser(_ArgumentParser):
             help="ignore args prefixed with an asterisk",
         )
         self.add_argument(
+            "-k",
+            "--ignore-kwargs",
+            action="store_true",
+            help="ignore kwargs prefixed with two asterisks",
+        )
+        self.add_argument(
             "-n", "--no-ansi", action="store_true", help="disable ansi output"
         )
         self.add_argument(