+++ tests/_test.py
@@ -845,3 +845,43 @@ def test_ignore_args(
         or name.startswith(PASS)
         and "w-args" in name
     )
+
+
+@pytest.mark.parametrize(
+    [NAME, TEMPLATE, "_"],
+    templates.registered.filtergroup(MULTI),
+    ids=templates.registered.filtergroup(MULTI).getids(),
+)
+def test_ignore_kwargs(
+    init_file: InitFileFixtureType,
+    main: MockMainType,
+    name: str,
+    template: str,
+    _: str,
+) -> None:
+    """Test docstrings without kwargs don't fail wih ``-k``.
+
+    Passing tests will fail and failing tests will pass, as tests which
+    generally pass will have kwargs documented, which shouldn't be with
+    this argument.
+
+    :param init_file: Initialize a test file.
+    :param main: Mock ``main`` function.
+    :param name: Name of test.
+    :param template: Contents to write to file.
+    """
+    file = init_file(template)
+    assert main(
+        long.check_class,
+        long.check_protected,
+        long.check_overridden,
+        long.check_dunders,
+        long.check_property_returns,
+        long.ignore_kwargs,
+        file.parent,
+    ) == int(
+        name.startswith(FAIL)
+        and "w-kwargs" not in name
+        or name.startswith(PASS)
+        and "w-kwargs" in name
+    )