+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [Unreleased](https://github.com/jshwi/docsig/compare/v0.32.0...HEAD)
 ------------------------------------------------------------------------
 ### Added
+- Add `-k/--ignore-kwargs` argument
 - Add `-a/--ignore-args` argument
 
 [0.32.0](https://github.com/jshwi/docsig/releases/tag/v0.32.0) - 2022-12-11